After numpy 1.24, np.float and np.int are no longer allowed. Using now int32 and float32 consistently.

+++ opendrift/elements/elements.py
@@ -62,7 +62,7 @@ class LagrangianArray:
                          'units': 's',
                          'seed': False,
                          'default': 0}),
-        ('origin_marker', {'dtype': np.int16,
+        ('origin_marker', {'dtype': np.int32,
                            'unit': '',
             'description': 'An integer kept constant during the simulation. Different values may be used for different seedings, to separate elements during analysis. With GUI, only a single seeding is possible.',
                            'default': 0}),
